Silence compiler warning
test/generate_pi.cpp:72:17: error:
      comparison of integers of different signs: 'value_type' (aka
'int') and
      'uint64_t' (aka 'unsigned long') [-Werror,-Wsign-compare]
    check(pi[n] == primesieve::count_primes(0, n));

+++ test/generate_pi.cpp
@@ -69,7 +69,7 @@ int main()
   {
     int n = dist(gen) % pi.size();
     cout << "pi(" << n << ") = " << pi[n];
-    check(pi[n] == primesieve::count_primes(0, n));
+    check(pi[n] == (int) primesieve::count_primes(0, n));
   }
 
   cout << endl;
